+++ src/components/table/table.tsx
@@ -771,7 +771,7 @@ export class Table {
             >
                 <limel-checkbox
                     onChange={this.selectAllOnChange}
-                    disabled={!this.totalRows}
+                    disabled={!this.data.length}
                     checked={this.tableSelection?.hasSelection}
                     indeterminate={
                         this.tableSelection?.hasSelection &&